setting up conftest for catalogs backends too
